+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - "Show elapsed time" checkbox that toggles the timestamp for a display mode
+- Text field length validation
 
 ### Changed